Updated Screen ratio
Updated logic for stretching. It will appear with a lot of white space
on a desktop, but this is where we'd use the background to make it look
so not out of place. We'd also probably cap the max height. On mobile it
shall scale and place bars to make it look correct. Viewing it on 480 x
800 or any integer multiple of that will show a filled screen.